Earlier prefix compression of vocabulary during index build (#949)
So far, the prefix compression of the internal vocabulary was done after the conversion of the triples from local `Id`s to global `Id`s. At that stage the sorting of the permutation (vis STXLL) already started. Both the prefix compression and that sorting require significant amounts of RAM, which led to unnecessary OOM errors when RAM was scarce. Now the prefix compression is performed before the sorting of the permutation starts.