Only allow const access to the `IndexMetaData` once it has been created (#947)
The `IndexMetaData` stores the `CompressedRelationMetaData` and the `CompressedBlockMetaData` for each permutation. With the changes from this commit, this data structure is created in a single pass by adding all the meta data in the correct order. After it has been created, only const access is allowed. This is an important preparation for compressing the `IndexMetaData` for the `Oxx` and `Sxx` permutations.
This commit also cleans up the interface of the `CompressedRelationWriter` to make it easier and safer to use.