Guard against timing attacks (#959)
When comparing a user-provided string against a stored secret via a trivial comparison, it is possible to incrementally figure out the secret by measuring the time until the comparison returns false. This is because typical 
implementations of string comparison immediately return after the first mismatching character. This commit implements a comparison of strings that is guaranteed to take linear time in the length of the compared strings and thus guards against such attacks. This mechanism is then used for the access token that allows for privileged operations on a server, like clearing the cache.