Refactoring the text index, Part 1 of N (#955)
In the text index so far, a lot of functions in the text index often accepted or returned a vector of Entities/Words, a vector of scores, and a vector of text records. These are now packed into a common struct which already makes the text index much more readable.