Fix CMake compiler compatibility error  (#988)
QLever requires G++ >= 11 or Clang++>=16 to succesfully compile. This was previously already checked in the `CMakeLists.txt`, but the failure of the check didn't lead to a failure of the CMake process, but only to a printed warning, and in consequence to much harder to read error messages when an unsupported compiler was used. 
This is now fixed and CMake fails early with a clear message stated that the used compiler is not supported.