Fix ORDER BY when the sort columns contain NaN values (#987)
So far, the behavior of an ORDER BY operation was undefined when NaN values were present in one of the ORDER BY columns. It could even lead to a crash of the server. Now the order is well-defined such that the NaN values always come after all other values when sorting in ascending order and before all other values when sorting in descending order.