Folding date values into the `Id` (#983)
So far, dates were represented via strings such that the lexicographical order of these strings corresponds to the chronological order of the corresponding dates. This kind of representation dates back to the early days of QLever, where it was also used for integer and double values. It is now no longer used within QLever for anything and all the associated code has been deleted.

A date is now encoded directly in the `Id`, just like `xsd:int` or `xsd:double` values already were. In particular, this allows a much more efficient computation of functions like YEAR, MONTH or DAY. Adherence to the standard has also been improved. In particular, more date types are supported now (namely: `xsd:dateTime`, `xsd:date`, `xsd:gYear`, and `xsd:gYearMonth`) and there is a proper encoding of the timezone information.

When the year is larger than 9999 or smaller than -9999, then only the year is encoded and no information about the month, day, hour, minute, second, or timezone. This allows the encoding of very large years, in particular, those pertaining to the early days of our universe. There are LOG(DEBUG) messages during parsing when information from the input is discarded. There is also a LOG(DEBUG) message when an invalid date is encountered: in that case, the date is now stored as an ordinary literal without type. In all other cases, a date in the result is now exactly equal to how that date was written in the input (which wasn't the case before).