Don't read index scans during query planning (#1007)
Previously, the query planner already performed index scans with two given elements (for example `?x <is-a> <Scientist>`). The rationale behind this decision was that these index scans have to be performed sooner or later anyway and that there is no choice in the permutation, because there is only one degree of freedom. Additionally performing the scans helps the query planner because it gets an exact size estimate for these index scans.
This commit changes this behavior: The query planner now never performs any index scans. We however still get the exact size for all index scans. This means that for index scans with two given elements, the first and the last block of the result have to be read from disk.

This commit also gets rid of a lot of code that was required to obtain the correct runtime information in the case that a query contained index scans that were already computed during the query planning.