Ordered threadsafe queue (#1011)
Add a class `OrderedThreadsafeQueue`, where a number of threads can produce values, where each value is associated with an index and the values are consumed strictly in the order of the indices. Each thread must produce its values in order of ascending indices and the union of these indices, over all threads, must form a consecutive sequence (without gaps) starting from 0, otherwise there will be a deadlock.