Enhance `Operation::getResult` by option `onlyReadFromCache` (#1013)
So far, `Operation::getResult` either gets the result from the cache (if it has already been computed and then cached before) or computes it from scratch. Now there is also an option to just get it from the cache (if it is there) or otherwise return `nullptr`. This is preparation for the "lazy joins" in the case where one of the join operands is an index scan: then we should proceed differently, depending on whether the result of the full index scan happens to be in the cache or not.