Extend `cppcoro::generator` by optional `Details` (#1014)
A `cppcoro::generator` can now have a `Details` object, which can be read and written at any point during generation. The generator can have its own `Details` object or a pointer to an object managed elsewhere. If no such details are needed, there is zero overhead compared to the implementation so far.