Add a helper function that reads and parses a JSON file (#1024)
A typical source of JSON data structures are files that end on `.json` and contain a single JSON object (think of a configuration file). This commit adds a function that takes a filename that ends on `.json` and parses its content as JSON. All possible errors (file can't be opened, parsing errors, etc.) are reported to the user in a way that is also understandable for end users.