More correct SPARQL Expressions using `ValueId`s and "ternary Boolean" semantics (#1038)
When evaluating SPARQL expressions, all numerical values are now represented as `ValueId`s. When computing with these `ValueId`s, the datatype is properly considered. For example, the sum of two `xsd:int` now remains `xsd:int`, whereas so far all expressions involving numerical values became `xsd:decimal`. The additional checking required has a performance overhead of around 50% compared to the previous (incorrect) code. This can be alleviated (in a future PR) by detecting the frequent special case, where all or most of the values in a column have the same datatype.

The [Effective Boolean Value](https://www.w3.org/TR/sparql11-query/#ebv) is now also implemented properly, as well as the "ternary" semantics when evaluating logical expressions involving `UNDEF` values. For example, `true || UNDEF` is `true`, but `false || UNDEF` is `UNDEF` and not `false`. This semantics is relevant for the evaluation of `||`, `&&`, `!`, and `FILTER`.