Only one `RAND()` per group in a GROUP BY (#1050)
Special treatment of `RAND()` when it is part of a `GROUP BY`, notably in a query of the form `... GROUP BY <some variables> ORDER BY RAND()`. We then expect a random order of the groups, and a new random order each time we repeat the query. In the code so far, this gave an error message. For Blazegraph, the order remains the same when repeating the query. This fixes #1049.