Fix index-version check (was wrong when e2e data changed) (#1051)
In the workflow so far, there was a false "new index version" error when a PR fixed a bug in the code (without changing the index version) and the test was also wrong, so that the bug did not lead to a failure in the e2e test. This is now fixed.