Make the exceptions generated by ANTLRErrorHandling more general (#1044)
We use ANTLR based parsers for multiple grammars (SPARQL, HTTP media types, configuration shorthands, ...). We have several utilities for ANTLR for the error handling. So far, the messages of some of the exceptions that were generated by these utils were hardcoded to the SPARQL grammar, so it could happen that an `InvalidSparqlQueryException` was raised when parsing a media type etc. This commit fixes these inconsistencies and also cleans up the structure of the ANTLR utilities.