A typesafe proxy class for the type-erased `ConfigOptions` (#1055)
Each `ConfigOption` is currently stored as a `std::variant` of all the supported types (This might be changed to a fully type-erased implementation in the future). This means, that we cannot ask a `ConfigOption` which type it holds at compile time, although we have to specify this type as a template argument during the construction of the `ConfigOption`. This commit adds a templated proxy class that stores a `ConfigOption` together with its value type. This proxy is returned as the result of `ConfigManager::addConfigOption` and allows for more typesafe manipulation of the `ConfigOption` after its creation.