Fix the Codecov toolchain (#1064)
Recently, the Codecov toolchain used to run into out-of-memory problems on the GitHub actions runners. In addition, the thread sanitizer and undefined behavior sanitizer runs took much longer than previously. I suspect the reason to be changes in a new Clang or LLVM version.
Both of these issues are now fixed by 
* disabling expensive tests for the expensive sanitizers (UB and thread sanitizer)
* Let the Codecov toolchain use the normal standard library on linux (`libstdc++`). It was previously using `libc++` but that is not necessary anymore, because we in the meantime have a MacOS build that uses `libc++`.