Implement CONCAT (#1074)
This is our second function with a variable number of arguments, after `COALESCE`. A non-trivial aspect of the implementation is the handling of repeated evaluations of the same `LiteralExpression`, like the `" : "` in `SELECT (CONCAT(?number, " : ", GROUP_CONCAT(?value)) AS ?values) WHERE { ?number wdt:P1181 ?value } GROUP BY ?number ORDER BY DESC(COUNT(?value))`. To make this more efficient, the `LiteralExpression::evaluate` method now implements a simple result cache. To make this cache safe under multi-threaded access (which is currently not needed, but could happen in the future), `std::atomic` is used, which has a small performance overhead. Eventually, we should find a more principled solution without that overhead.