Sub managers for the `ConfigOptionManager` class (#1071)
Add the posibility to add a ConfigManager as a child to another ConfigManager. This makes it easier to manage configurations that have a nested structure. For example, the configuration for QLever's index builder takes settings for the Locale. These locale settings consist of multiple subsettings (the language, the country, etc.).