Make `MemorySize` factory work with `int` arguments. (#1094)
Before, a call of MemorySize::kilobytes(4) was ambiguous, because the factory function overloads took size_t or double, to both of which an int can be implicitly converted. The same was true for similar calls to megabytes(), gigabytes() etc. This has been fixed by templating these functions for arbitrary numeric types.