Add a Websocket server that allows for interaction with running queries (#970)
When posting a query to QLever via a blocking HTTP request, it is now possible to assign a unique ID to this query. Using this ID it is possible to also open a Websocket connection at the same time that will receive status updates on the running query. Currently, the partial runtime information trees are sent after each operation has finished. In the future, we will always get complete trees, and the websocket mechanism can also be used to interact with the query, for example to cancel it.