Fix deadlocks in the Websocket module caused by heavy load (#1109)
The destructor of a query session has to synchronously wait until the session is unregistered. This waiting is blocking because destructors can't be asynchronous in C++. Under heavy load it could happen that all available server threads were blocking in such destructors, which lead to a deadlock as there were no threads left which could complete the tasks that the destructors were blocking on.
This is now fixed as the destructors are also continuing to run server tasks while they are waiting for the unregistering of their session.

Also clear up the interface of the QueryHub class. It is now only possible to have at most one sender per query session, else an exception is thrown. This makes reasoning about the lifetime of such sessions much more explicit.