Better move semantics for `IdTable`s (#1110)
When an `IdTable` is moved from, it now becomes a valid `IdTable` with the same number of columns as before the move and zero rows. So far, it became an invalid `IdTable` with no columns. This simplifies the code in various places. In particular, we can now simple call `clear()` on an `IdTable` in order to empty it.

This comes with a small overhead, namely for allocating space for `k` new empty column vectors after the move, where `k` is the number of columns before the move. However, this should not matter in practice because we handle only few `IdTable`s, which are either small (in which case everything is very fast anyway) or they have a large number of rows (in which case the cost for creating `k` empty colums is negligible). Should be performance overhead really matter at some point, we could implement a `small_vector` type, that allocates storage for the columns on the stack when the number of columns is below a certain threshold.