Decrease the block size of `CompressedExternalIdTable` to 500 kB (#1115)
The block size for a single column of a compressed `IdTable` used to be `4 MB` and is now `500 kB`. This is still large enough so that a single block is well compressible and the space for the metadata is negligible compared to the space for the compressed blocks.

The index build needs RAM for `6 k` such blocks when sorting the permutations, where `k` is the number of runs of the merge sort. For an index build of OSM planet with 44 B input triples, the merge sort had 421 runs, which with a block size of `4 MB` leads to a RAM requirement of `10.1 GB`. With the new block size of `500 kB`, one fourth of this is sufficient.