Add explicit seeds to the random number generators (#1126)
For the pseudorandom number generators in util/Random.h, add the possibility to explicitly specify the initial random seed. This is useful for deterministic unit tests and benchmark where we want to reproduce random numbers. The default value for the random seed is always generated via std::default_random_engine to get as close to "true" randomness as possible in the default case.
Also move the functions and types inside the Random.h file to the ad_utility namespace.