Generalize and refactor query cancellation mechanism (#1125)
Generalize QLever's original timeout mechanism to a more general cancellation mechanism, which allows a query to be canceled either due to a timeout or due to a manual cancellation of the query. In the process, the timeout mechanism was also refactored: The timers used to manage the timeout mechanism are now implemented asynchronously using Boost::Asio. This means that there is a single thread for the whole server that will cancel queries as they time out.
In a following PR a mechanism that actually allows for queries to be canceled via the HTTP or Websocket API will be implemented and also integrated into the QLeverUI.