Fix bugs regarding query cancellation and server threads (#1149)
1. Fix a bug in the recursive query cancellation (introduced by #1125), which led to enormous runtimes for queries with many operations, like https://qlever.cs.uni-freiburg.de/wikidata/yI0y30 or https://qlever.cs.uni-freiburg.de/osm-planet/8Cn1w5 .

2. Fix a bug in the management of multiple server threads (introduced by #1103), which led to the server effectively not being multi-threaded anymore.

3. Make the thread sanitizer use Clang 17 (with CLang 16 there was a bug when a strand was resumed on another thread).

NOTE: The combination of 1+2 made QLever very unstable, since any query with many operations would stall the server. As a side effect, our proxy web server (which currently has a large timeout) was also affected because of unprocessed requests piling up. This was a nerve-racking but very interesting lesson in many respects.