Consistently replace `boost::asio::dispatch` by `boost::asio::post` (#1152)
`dispatch` and `post` have a different behavior when being called from a `strand`. `post` consistently gives up the current strand while `dispatch` keeps it. This might lead to deadlocks when the handler that is `dispatch`ed tries to acquire the same strand again.
Also extend the unit tests for the asynchronous websocket module s.t. they cover more of these subtleties.
See also #1149 which applied a similar fix (but not consistently across the codebase) as part of an urgent bugfix.