Print `ConfigOptions` in the order they were specified (#1153)
When `ConfigOptions` are printed (e.g. in a help message or to print the current configuration), they are now consistently printed in the order in which they were added to  the corresponding `ConfigManager`. This required non-trivial changes to the code, as conceptually these options are stored in a hash map to access them by their name. These hash maps now also store the order in which the entries were added.